feat(example): Add a simple game that the brain can play
In the game a player is getting hungry over time.
The player will die when it is too hungry.
The player can eat plants when it is in the same location.
The player can move over the field.
Moving the player happens every frametick in the chosen direction.
The brain can choose the direction through actors.
Sensors (moving closer to a plant, and standing still) incite the brain.
So far, the brain's performance in the game is relative to its random generation.
The brain does not learn. The brain does not change.
Only the input for the brain changes due to the sensors.